requirements.txt
torch==1.7.1
numpy==1.19.5
pylru==1.2.0
python-chess==1.999
protobuf==3.14.0
six==1.15.0